LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-window-open-expected.txt

CONSOLE MESSAGE: line 49: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
Opener Frame

Test
